# docs/rest/REST-API-OPENAPI.yaml
openapi: 3.1.0
info:
  title: eMuleBB REST API (canonical superset, emulebb-rust)
  version: 1.0.0
  x-contract-version: "1.0.0"
  x-contract-baseline: "eMuleBB 0.7.3 contract"
  summary: The one capability-gated /api/v1 contract; emulebb-rust owns the superset.
  description: >
    Source-of-truth for the single capability-gated `/api/v1` contract. There is
    one contract; implementations differ by the capabilities they advertise, not
    by a separate spec. emulebb-rust is the superset (it leads the contract
    version and adds indexing/Arr capabilities); the frozen eMuleBB MFC client is
    a subset advertising fewer capabilities. It is baselined on the eMuleBB
    `0.7.3` contract and evolves on its own independent **contract version** line
    (`x-contract-version`, semver), decoupled from any product release tag.
    Controllers (TrackMuleBB) read `GET /api/v1/capabilities`, target a contract
    version range, and call only advertised operations.
    Versioning policy: additive endpoints/fields bump the minor; breaking changes
    bump the major with a documented migration. Capability model:
    `emulebb-tooling/docs/active/API-V1-COMPATIBILITY.md`.
    The API uses `X-API-Key` authentication only; adapter compatibility
    (`/api/v2` qBit-compat, Torznab) must not weaken the native route shape.
servers:
  - url: http://{host}:{port}/api/v1
    variables:
      host:
        default: 127.0.0.1
      port:
        default: "4711"
...
    variables:
      host:
...
security:
  - ApiKeyAuth: []
...
paths:
  /app:
    get:
      tags: [App]
      summary: Return application identity, runtime state, and static capabilities.
      operationId: getApp
      responses:
        "200":
          $ref: "#/components/responses/AppResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "405":
          $ref: "#/components/responses/MethodNotAllowedResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
        "503":
          $ref: "#/components/responses/ErrorResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /capabilities:
    get:
      tags: [App]
      summary: Return contract version and advertised capability tokens.
      operationId: getCapabilities
      responses:
        "200":
          $ref: "#/components/responses/CapabilityDiscoveryResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "405":
          $ref: "#/components/responses/MethodNotAllowedResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
        "503":
          $ref: "#/components/responses/ErrorResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /app/preferences:
    get:
      tags: [App]
      summary: Return controller-relevant preferences.
      operationId: getPreferences
      responses:
        "200":
          $ref: "#/components/responses/PreferencesResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "405":
          $ref: "#/components/responses/MethodNotAllowedResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
        "503":
          $ref: "#/components/responses/ErrorResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
    patch:
      tags: [App]
      summary: Update controller-relevant preferences.
      operationId: patchPreferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PreferencesPatch"
      responses:
        "200":
          $ref: "#/components/responses/PreferencesResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "405":
          $ref: "#/components/responses/MethodNotAllowedResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
        "503":
          $ref: "#/components/responses/ErrorResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /app/shutdown:
    post:
      tags: [App]
      summary: Request clean eMule process shutdown.
      operationId: shutdownApp
      description: >
        This endpoint closes eMule only. Operating-system shutdown and reboot
        actions from the legacy WebServer are intentionally not part of the
        REST API.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShutdownRequest"
      responses:
        "200":
          $ref: "#/components/responses/OkAcceptedResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "405":
          $ref: "#/components/responses/MethodNotAllowedResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
        "503":
          $ref: "#/components/responses/ErrorResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /diagnostics/dumps:
    post:
      tags: [Diagnostics]
      summary: Capture an operator-requested eMule diagnostic dump.
      operationId: captureDiagnosticDump
      description: >
        Unsafe diagnostic endpoint. It is available only when diagnostic REST
        endpoints are explicitly enabled in the WebServer configuration.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiagnosticDumpRequest"
      responses:
        "200":
          $ref: "#/components/responses/DiagnosticDumpResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "405":
          $ref: "#/components/responses/MethodNotAllowedResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
        "503":
          $ref: "#/components/responses/ErrorResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /diagnostics/crash-tests:
    post:
      tags: [Diagnostics]
      summary: Trigger a diagnostic crash to verify dump generation.
      operationId: triggerDiagnosticCrashTest
      description: >
        Unsafe diagnostic endpoint. It is available only when diagnostic REST
        endpoints are explicitly enabled in the WebServer configuration. A
        successful request intentionally terminates the process, so clients
        should treat connection loss after acceptance as the expected result.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiagnosticCrashTestRequest"
      responses:
        "200":
          $ref: "#/components/responses/OkAcceptedResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "405":
          $ref: "#/components/responses/MethodNotAllowedResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
        "503":
          $ref: "#/components/responses/ErrorResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /status:
    get:
      tags: [Stats]
      summary: Return current network, transfer, server, and Kad status.
      operationId: getStatus
      responses:
        "200":
          $ref: "#/components/responses/StatusResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "405":
          $ref: "#/components/responses/MethodNotAllowedResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
        "503":
          $ref: "#/components/responses/ErrorResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /snapshot:
    get:
      tags: [Stats]
      summary: Return a controller dashboard snapshot.
      operationId: getSnapshot
      parameters:
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          $ref: "#/components/responses/SnapshotResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "405":
          $ref: "#/components/responses/MethodNotAllowedResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
        "503":
          $ref: "#/components/responses/ErrorResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /stats:
    get:
      tags: [Stats]
      summary: Return detailed cumulative and session statistics.
      operationId: getStats
      responses:
        "200":
          $ref: "#/components/responses/StatsResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "405":
          $ref: "#/components/responses/MethodNotAllowedResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
        "503":
          $ref: "#/components/responses/ErrorResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /categories:
    get:
      tags: [Categories]
      summary: List download categories.
      operationId: listCategories
      responses:
        "200":
          $ref: "#/components/responses/CategoryListResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "405":
          $ref: "#/components/responses/MethodNotAllowedResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
        "503":
          $ref: "#/components/responses/ErrorResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
    post:
      tags: [Categories]
      summary: Create a download category.
      operationId: createCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryCreateRequest"
      responses:
        "200":
          $ref: "#/components/responses/CategoryResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "405":
          $ref: "#/components/responses/MethodNotAllowedResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
        "503":
          $ref: "#/components/responses/ErrorResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /categories/{categoryId}:
    parameters:
      - $ref: "#/components/parameters/CategoryId"
    get:
      tags: [Categories]
      summary: Return one download category.
      operationId: getCategory
      responses:
        "200":
          $ref: "#/components/responses/CategoryResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "405":
          $ref: "#/components/responses/MethodNotAllowedResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
        "503":
          $ref: "#/components/responses/ErrorResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
    patch:
      tags: [Categories]
      summary: Update one download category.
      operationId: patchCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryPatch"
      responses:
        "200":
          $ref: "#/components/responses/CategoryResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "405":
          $ref: "#/components/responses/MethodNotAllowedResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
        "503":
          $ref: "#/components/responses/ErrorResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
    delete:
      tags: [Categories]
      summary: Delete one download category.
      operationId: deleteCategory
      responses:
        "200":
          $ref: "#/components/responses/OkResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "405":
          $ref: "#/components/responses/MethodNotAllowedResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
        "503":
          $ref: "#/components/responses/ErrorResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /transfers:
    get:
      tags: [Transfers]
      summary: List incomplete and completed download transfers.
      operationId: listTransfers
      parameters:
        - name: state
          in: query
          schema:
            $ref: "#/components/schemas/TransferState"
        - name: categoryId
          in: query
          schema:
            type: integer
            minimum: 0
            maximum: 4294967295
        - $ref: "#/components/parameters/Offset"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          $ref: "#/components/responses/TransferListResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "405":
          $ref: "#/components/responses/MethodNotAllowedResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
        "503":
          $ref: "#/components/responses/ErrorResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
    post:
      tags: [Transfers]
      summary: Add one or more ED2K links to downloads.
      operationId: createTransfers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferCreateRequest"
      responses:
        "200":
          $ref: "#/components/responses/BulkOperationResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "405":
          $ref: "#/components/responses/MethodNotAllowedResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
        "503":
          $ref: "#/components/responses/ErrorResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /transfers/operations/clear-completed:
    post:
      tags: [Transfers]
      summary: Clear completed transfers from the transfer list.
      operationId: clearCompletedTransfers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClearCompletedTransfersRequest"
      responses:
        "200":
          $ref: "#/components/responses/OkResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "405":
          $ref: "#/components/responses/MethodNotAllowedResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
        "503":
          $ref: "#/components/responses/ErrorResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /transfers/{hash}:
    parameters:
      - $ref: "#/components/parameters/FileHash"
    get:
      tags: [Transfers]
      summary: Return one transfer row.
      operationId: getTransfer
      responses:
        "200":
          $ref: "#/components/responses/TransferResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "405":
          $ref: "#/components/responses/MethodNotAllowedResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
        "503":
          $ref: "#/components/responses/ErrorResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
    patch:
      tags: [Transfers]
      summary: Update mutable transfer properties.
      operationId: patchTransfer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferPatch"
      responses:
        "200":
          $ref: "#/components/responses/TransferResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "405":
          $ref: "#/components/responses/MethodNotAllowedResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
        "503":
          $ref: "#/components/responses/ErrorResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
    delete:
      tags: [Transfers]
      summary: Remove a completed transfer row without deleting or unsharing the completed file.
      description: >
        Removes a completed transfer from the native transfer list without
        deleting the completed file or changing its shared-file registration.
        Incomplete transfers cannot preserve their partial .part state when cancelled; use
        `DELETE /transfers/{hash}/files?confirm=true` to cancel an incomplete
        transfer and delete its local part data.
      operationId: deleteTransfer
      responses:
        "200":
          $ref: "#/components/responses/BulkOperationResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "405":
          $ref: "#/components/responses/MethodNotAllowedResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
        "503":
          $ref: "#/components/responses/ErrorResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /transfers/{hash}/files:
    parameters:
      - $ref: "#/components/parameters/FileHash"
    delete:
      tags: [Transfers]
      summary: Delete one transfer and its local data.
      description: >
        Cancels/removes the transfer and deletes the local completed file or
        incomplete .part data. This route is the REST v1 replacement for the
        pre-RC body-required destructive transfer delete payload.
      operationId: deleteTransferFiles
      parameters:
        - $ref: "#/components/parameters/Confirm"
      responses:
        "200":
          $ref: "#/components/responses/BulkOperationResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "405":
          $ref: "#/components/responses/MethodNotAllowedResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
        "503":
          $ref: "#/components/responses/ErrorResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /transfers/{hash}/details:
    parameters:
      - $ref: "#/components/parameters/FileHash"
    get:
      tags: [Transfers]
      summary: Return detailed part and source state for one transfer.
      operationId: getTransferDetails
      responses:
        "200":
          $ref: "#/components/responses/TransferDetailsResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "405":
          $ref: "#/components/responses/MethodNotAllowedResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
        "503":
          $ref: "#/components/responses/ErrorResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /transfers/{hash}/sources:
    parameters:
      - $ref: "#/components/parameters/FileHash"
    get:
      tags: [Transfers]
      summary: Return known sources for one transfer.
      operationId: listTransferSources
      responses:
        "200":
          $ref: "#/components/responses/TransferSourceListResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "405":
          $ref: "#/components/responses/MethodNotAllowedResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
        "503":
          $ref: "#/components/responses/ErrorResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...
  responses:
    AppResponse:
      description: Application envelope.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AppEnvelope"
...
    CapabilityDiscoveryResponse:
      description: Capability discovery envelope.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/CapabilityDiscoveryEnvelope"
...
    ErrorResponse:
      description: Error envelope.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...